Create SAS.Project2
This project explores the leading causes of death in the United States, focusing on the relationship between different age groups and their susceptibility to various causes. Using a dataset from data.gov, the project leverages SAS to perform in-depth analysis and visualization.

The dataset is first imported and cleaned, removing summary records for "All Causes" to concentrate on specific causes of death.  A new variable, "Age Category," is created to categorize individuals based on age-adjusted death rates, enhancing the analysis.  Descriptive labels are added to variables for clarity.

The analysis then dives into understanding the distribution of deaths across age categories and causes. A contingency table is generated to examine the relationship between age categories and causes of death, providing insights into the most common causes for each age group.  The mean and standard deviation of age-adjusted death rates are calculated for different age categories, revealing significant variations in mortality rates.